Split asset manager controller in files and transform to async/await
